fix: CLI 追帧循环改 last_tick+=interval 防丢 tick；ClearResult u64 适配

+++ crates/tetris-cli/src/app.rs
@@ -22,7 +22,7 @@ pub enum AppState {
         clear_flash_timer: u8,
         score_flash_timer: u8,
         prev_grid: [[CellType; 10]; 20],
-        prev_flash_mask: u32,
+        prev_flash_mask: u64,
         prev_half: bool,
     },
     Pause {
@@ -52,7 +52,7 @@ pub enum AppState {
         clear_flash_timer: u8,
         score_flash_timer: u8,
         prev_grid: [[CellType; 10]; 20],
-        prev_flash_mask: u32,
+        prev_flash_mask: u64,
         prev_half: bool,
         spectating: Option<usize>,
     },
